Add method to save computer
